pyproject.toml
[tool.poetry]
name = "mushysoft_admin"
version = "0.1.0"
description = "A lightweight and flexible admin panel for FastAPI"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/your_username/mushysoft_admin"
homepage = "https://github.com/your_username/mushysoft_admin"
keywords = ["FastAPI", "Admin Panel", "SQLAlchemy", "Authentication"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
asyncpg = "^0.30.0"
pydantic = "^2.6.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"
httpx = "^0.28.1"
jinja2 = "^3.1.3"
pydantic-settings = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
